pyproject.toml
[tool.poetry]
name = "steam-screenshot-reporter"
version = "0.1.0"
description = ""
authors = ["Lauri Elias <lauri@indoorsman.ee>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
beautifulsoup4 = "^4.10.0"
Django = "^4.0.1"
psycopg2 = "^2.9.3"
uWSGI = "^2.0.20"
NudeNet = "^2.0.9"
Pillow = "^9.0.0"
numpy = "^1.22.1"
django-dotenv = "^1.4.2"

[tool.poetry.dev-dependencies]
pytest-cov = "^3.0.0"
pre-commit = "^2.16.0"
flake8 = "^4.0.1"
mypy = "^0.931"
isort = "^5.10.1"
black = {version = "^21.12b0", allow-prereleases = true}
types-requests = "^2.27.7"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.black]
line-length = 79
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
